+++ src/features/share/lib/initKakaoSDK.ts
@@ -1,0 +1,36 @@
+const KAKAO_SDK = {
+  src: 'https://t1.kakaocdn.net/kakao_js_sdk/2.7.4/kakao.min.js',
+  integrity:
+    'sha384-DKYJZ8NLiK8MN4/C5P2dtSmLQ4KwPaoqAfyA/DfmEc1VDxu4yyC7wy6K1Hs90nka',
+  crossorigin: 'anonymous',
+};
+
+let kakaoInitPromise: Promise<void> | null = null;
+
+const initKakaoSDK = (): Promise<void> => {
+  // 1. 이미 초기화된 경우
+  if (window.Kakao?.isInitialized()) return Promise.resolve();
+  // 2. 로딩 중이거나 이미 로드된 경우 — 같은 Promise 반환 (로딩 중 중복 로딩을 막기 위함)
+  if (kakaoInitPromise) return kakaoInitPromise;
+  // 3. 스크립트 로드 시작
+  kakaoInitPromise = new Promise((resolve, reject) => {
+    const kakaoSdk = document.createElement('script');
+    kakaoSdk.id = 'kakao-sdk';
+    kakaoSdk.src = KAKAO_SDK.src;
+    kakaoSdk.integrity = KAKAO_SDK.integrity;
+    kakaoSdk.crossOrigin = KAKAO_SDK.crossorigin;
+    document.head.appendChild(kakaoSdk);
+    kakaoSdk.onload = () => {
+      window.Kakao.init(import.meta.env.VITE_KAKAO_JAVASCRIPT_KEY);
+      resolve();
+    };
+    kakaoSdk.onerror = () => {
+      kakaoSdk.remove();
+      kakaoInitPromise = null;
+      reject(new Error('Kakao SDK 로드 실패'));
+    };
+  });
+  return kakaoInitPromise;
+};
+
+export default initKakaoSDK;
