+++ src/features/share/lib/shareFormat.ts
@@ -1,3 +1,5 @@
+import type { ShareData } from '../model/share.type';
+
 const TITLE = '즐거운 만남 끝까지 즐겁게, 모또';
 const DESCRIPTION = '정산을 요청했어요! 금액을 확인하고 정산해주세요.';
