+++ src/features/share/model/share.type.ts
